Move all logic of computer setup into its own class
